chore: move module version resolution tests to garden tests
the logic for the resolution is in garden class now so it
makes sense to have the tests there as well.
