+++ core/test/unit/src/vcs/vcs.ts
@@ -17,8 +17,6 @@ import {
   VcsFile,
   getModuleTreeCacheKey,
   hashModuleVersion,
-  ModuleVersions,
-  ModuleVersion,
   NamedModuleVersion,
   NamedTreeVersion,
 } from "../../../../src/vcs/vcs"
@@ -34,12 +32,10 @@ import { realpath, readFile, writeFile } from "fs-extra"
 import { DEFAULT_API_VERSION, GARDEN_VERSIONFILE_NAME } from "../../../../src/constants"
 import { defaultDotIgnoreFiles, fixedProjectExcludes } from "../../../../src/util/fs"
 import { LogEntry } from "../../../../src/logger/log-entry"
-import { findByName } from "../../../../src/util/util"
 
-class TestVcsHandler extends VcsHandler {
+export class TestVcsHandler extends VcsHandler {
   name = "test"
   private testTreeVersions: TreeVersions = {}
-  private testModuleVersions: ModuleVersions = {}
 
   async getRepoRoot() {
     return "/foo"
@@ -81,12 +77,6 @@ class TestVcsHandler extends VcsHandler {
 describe("VcsHandler", () => {
   let handlerA: TestVcsHandler
   let gardenA: TestGarden
-
-  // note: module-a has a version file with this content
-  const treeVersionA: TreeVersion = {
-    contentHash: "1234567890",
-    files: [],
-  }
 
   beforeEach(async () => {
     gardenA = await makeTestGardenA()
@@ -268,105 +258,6 @@ describe("VcsHandler", () => {
 
       const result = await handlerA.resolveTreeVersion(gardenA.log, gardenA.projectName, moduleConfig)
       expect(result).to.eql(version)
-    })
-  })
-
-  describe("resolveModuleVersion", () => {
-    beforeEach(() => {
-      gardenA.clearCaches()
-    })
-
-    it("should return module version if there are no dependencies", async () => {
-      const module = await gardenA.resolveModule("module-a")
-      gardenA.vcs = handlerA
-      const result = await gardenA.resolveModuleVersion(gardenA.log, module, [])
-
-      expect(result).to.eql({
-        versionString: getModuleVersionString(module, { ...treeVersionA, name: "module-a" }, []),
-        dependencyVersions: {},
-        files: [],
-      })
-    })
-
-    it("should hash together the version of the module and all dependencies", async () => {
-      const moduleConfigs = await gardenA.resolveModules({
-        log: gardenA.log,
-      })
-      gardenA.vcs = handlerA
-
-      const moduleA = findByName(moduleConfigs, "module-a")!
-      const moduleB = findByName(moduleConfigs, "module-b")!
-      const moduleC = findByName(moduleConfigs, "module-c")!
-
-      gardenA.clearCaches()
-
-      const moduleVersionA: ModuleVersion = {
-        versionString: treeVersionA.contentHash,
-        files: [],
-        dependencyVersions: {},
-      }
-      moduleA.version = moduleVersionA
-      handlerA.setTestTreeVersion(moduleA.path, treeVersionA)
-
-      const versionStringB = "qwerty"
-      const moduleVersionB: ModuleVersion = {
-        versionString: versionStringB,
-        files: [],
-        dependencyVersions: { "module-a": moduleVersionA },
-      }
-      moduleB.version = moduleVersionB
-      const treeVersionB: TreeVersion = { contentHash: versionStringB, files: [] }
-      handlerA.setTestTreeVersion(moduleB.path, treeVersionB)
-
-      const versionStringC = "asdfgh"
-      const treeVersionC: TreeVersion = { contentHash: versionStringC, files: [] }
-      handlerA.setTestTreeVersion(moduleC.path, treeVersionC)
-
-      const gardenResolvedModuleVersion = await gardenA.resolveModuleVersion(gardenA.log, moduleC, [moduleA, moduleB])
-      const manuallyResolvedModuleVersion = {
-        versionString: getModuleVersionString(moduleC, { ...treeVersionC, name: "module-c" }, [
-          { ...moduleVersionA, name: "module-a" },
-          { ...moduleVersionB, name: "module-b" },
-        ]),
-        dependencyVersions: {
-          "module-a": moduleVersionA,
-          "module-b": moduleVersionB,
-        },
-        files: [],
-      }
-
-      expect(gardenResolvedModuleVersion).to.eql(manuallyResolvedModuleVersion)
-    })
-
-    it("should not include module's garden.yml in version file list", async () => {
-      const moduleConfig = await gardenA.resolveModule("module-a")
-      const version = await gardenA.resolveModuleVersion(gardenA.log, moduleConfig, [])
-      expect(version.files).to.not.include(moduleConfig.configPath!)
-    })
-
-    it("should be affected by changes to the module's config", async () => {
-      const moduleConfig = await gardenA.resolveModule("module-a")
-      const version1 = await gardenA.resolveModuleVersion(gardenA.log, moduleConfig, [])
-      moduleConfig.name = "foo"
-      const version2 = await gardenA.resolveModuleVersion(gardenA.log, moduleConfig, [])
-      expect(version1).to.not.eql(version2)
-    })
-
-    it("should not be affected by changes to the module's garden.yml that don't affect the module config", async () => {
-      const projectRoot = getDataDir("test-projects", "multiple-module-config")
-      const garden = await makeTestGarden(projectRoot)
-      const moduleConfigA1 = await garden.resolveModule("module-a1")
-      const configPath = moduleConfigA1.configPath!
-      const orgConfig = await readFile(configPath)
-
-      try {
-        const version1 = await gardenA.resolveModuleVersion(garden.log, moduleConfigA1, [])
-        await writeFile(configPath, orgConfig + "\n---")
-        const version2 = await gardenA.resolveModuleVersion(garden.log, moduleConfigA1, [])
-        expect(version1).to.eql(version2)
-      } finally {
-        await writeFile(configPath, orgConfig)
-      }
     })
   })
 })
